// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Profile {
  id: number;
  name: string;
  title: string;
  /**
   * One-line positioning statement for the hero
   */
  tagline: string;
  bio: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  location?: string | null;
  headshot?: (number | null) | Media;
  socialLinks?:
    | {
        platform: string;
        url: string;
        id?: string | null;
      }[]
    | null;
  resume?: (number | null) | Media;
  calendlyUrl?: string | null;
  email?: string | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}